Removed spurious dir level
