+++ demosauruswebapp/static/js/author-page.js
