+++ demosauruswebapp/templates/publication/subjects-tab.html
