+++ demosauruswebapp/templates/publication/view.html
